[runtime] 32-Bit Queue Descriptors

+++ src/rust/runtime/queue/qdesc.rs
@@ -7,7 +7,7 @@
 
 /// IO Queue Descriptor
 #[derive(Debug, Eq, PartialEq, Hash, Copy, Clone)]
-pub struct QDesc(usize);
+pub struct QDesc(u32);
 
 //==============================================================================
 // Trait Implementations
@@ -23,20 +23,20 @@ impl From<QDesc> for i32 {
 impl From<i32> for QDesc {
     /// Converts a [i32] to a [QDesc].
     fn from(val: i32) -> Self {
-        QDesc(val as usize)
+        QDesc(val as u32)
     }
 }
 
-impl From<QDesc> for usize {
-    /// Converts a [QDesc] to a [usize].
+impl From<QDesc> for u32 {
+    /// Converts a [QDesc] to a [u32].
     fn from(val: QDesc) -> Self {
         val.0
     }
 }
 
-impl From<usize> for QDesc {
-    /// Converts a [usize] to a [QDesc].
-    fn from(val: usize) -> Self {
+impl From<u32> for QDesc {
+    /// Converts a [u32] to a [QDesc].
+    fn from(val: u32) -> Self {
         QDesc(val)
     }
 }
